backend/src/requirements.txt
google-genai
boto3